pulled out code from ege tests into a base classes before separating the ege functions for metadata, annotations, relationships, etc.
